Wasabi

Documentation for Wasabi.

Wasabi.begin_transaction

Wasabi.colnames

Wasabi.coltype

Wasabi.commit

Wasabi.connect

Wasabi.create_schema

Wasabi.create_schema

Wasabi.delete_schema

Wasabi.delete_schema

Wasabi.df2model

Wasabi.disconnect

Wasabi.execute_raw_query

Wasabi.execute_raw_query

Wasabi.first

Wasabi.get_column_type

Wasabi.insert

Wasabi.isnullable

Wasabi.model2tuple

Wasabi.rollback

Wasabi.tablename

Wasabi.begin_transaction — Function

begin_transaction(db::Any)
Begins a transaction.

Wasabi.colnames — Method

colnames(m::Type{T}) where {T <: Model}
Returns the names of the columns for the given model.

Wasabi.coltype — Method

coltype(m::Type{T}, field::Symbol) where {T <: Model}
Returns the julia type of the given column.

Wasabi.commit — Function

commit(db::Any)
Commits the current transaction.

Wasabi.connect — Function

connect(config::ConnectionConfiguration)
Connects to the database using the given configuration.

Wasabi.create_schema — Function

create_schema(db::Any, m::Type{T}, constraints::Vector{ModelConstraint}) where {T <: Model}
Creates the schema for the given model and constraints.

Wasabi.create_schema — Method

create_schema(db::Any, m::Type{T}, constraints::Vector{ModelConstraint}) where {T <: Model}
Creates the schema for the given model.
Returns the query used to create the schema.

Wasabi.delete_schema — Function

delete_schema(db::Any, m::Type{T}) where {T <: Model}
Deletes the schema for the given model.

Wasabi.delete_schema — Method

delete_schema(db::Any, m::Type{T}) where {T <: Model}
Deletes the schema for the given model.

Wasabi.df2model — Method

df2model(m::Type{T}, df::DataFrame) where {T <: Model}
Converts the given DataFrame to the given model.

Wasabi.disconnect — Function

disconnect(db::Any)
Disconnects from the database.

Wasabi.execute_raw_query — Function

execute_raw_query(db::Any, query::String, params::Vector{Any})
Executes the given query with the given parameters.

Wasabi.execute_raw_query — Function

execute_raw_query(db::SQLite.DB, query::String, params::Vector{Any})
Executes the given query with the given parameters.

Wasabi.first — Function

first(db::Any, m::Type{T})
Returns the first row of the given model with the given id.

Wasabi.get_column_type — Method

get_column_type(col::Symbol, m::Type{T})::String where {T <: Model}
Returns the SQL type of the given column.

Wasabi.insert — Function

insert(db::Any, model::T) where {T <: Model}
Inserts the given model into the database.

Wasabi.isnullable — Method

isnullable(m::Type{T}, field::Symbol, constraints::Vector{S}) where {T <: Model,S<:Wasabi.ModelConstraint}
Returns true if the given column is nullable.

Wasabi.model2tuple — Method

model2tuple(m::T) where {T <: Model}
Converts the given model to a tuple.

Wasabi.rollback — Function

rollback(db::Any)
Rolls back the current transaction.

Wasabi.tablename — Method

tablename(m::Type{T}) where {T <: Model}
Returns the name of the table for the given model.