API

Wasabi.QueryBuilder.Join

Wasabi.QueryBuilder.Query

Wasabi.QueryBuilder.SelectExpr

Wasabi.Migrations.execute

Wasabi.Migrations.generate

Wasabi.Migrations.get_current_version

Wasabi.Migrations.get_last_version

Wasabi.Migrations.get_versions

Wasabi.Migrations.init

Wasabi.QueryBuilder.build

Wasabi.QueryBuilder.from

Wasabi.QueryBuilder.groupby

Wasabi.QueryBuilder.join

Wasabi.QueryBuilder.limit

Wasabi.QueryBuilder.offset

Wasabi.QueryBuilder.orderby

Wasabi.QueryBuilder.select

Wasabi.QueryBuilder.select

Wasabi.QueryBuilder.select

Wasabi.QueryBuilder.where

Wasabi.alias

Wasabi.alias

Wasabi.all

Wasabi.begin_transaction

Wasabi.colnames

Wasabi.coltype

Wasabi.commit!

Wasabi.connect

Wasabi.constraints

Wasabi.create_schema

Wasabi.delete!

Wasabi.delete_all!

Wasabi.delete_schema

Wasabi.df2model

Wasabi.disconnect

Wasabi.execute_query

Wasabi.first

Wasabi.foreign_keys

Wasabi.from_sql_value

Wasabi.insert!

Wasabi.isnullable

Wasabi.mapping

Wasabi.model2tuple

Wasabi.primary_key

Wasabi.rollback

Wasabi.tablename

Wasabi.tablename

Wasabi.to_sql_value

Wasabi.unique_constraints

Wasabi.update!

Wasabi.alias — Method

alias(m::String)
Returns the alias of the table for the given model.

Wasabi.alias — Method

alias(m::Type{T}) where {T <: Model}
Returns the alias of the table for the given model.

Wasabi.all — Function

all(db::Any, m::Type{T}) where {T <: Model}
Returns all rows of the given model.

Wasabi.begin_transaction — Function

begin_transaction(db::Any)
Begins a transaction.

Wasabi.colnames — Method

colnames(m::Type{T}) where {T <: Model}
Returns the names of the columns for the given model.

Wasabi.coltype — Method

coltype(t::Type{S}, col::Symbol, m::Type{T})::String where {T <: Model}
Returns the column type for the given column and model.

Wasabi.commit! — Function

commit!(db::Any)
Commits the current transaction.

Wasabi.connect — Function

connect(config::ConnectionConfiguration)
Connects to the database using the given configuration.

Wasabi.constraints — Method

constraints(m::Type{T}) where {T <: Model}
Returns the constraints for the given model.

Wasabi.create_schema — Function

create_schema(db::Any, m::Type{T}) where {T <: Model}
Creates the schema for the given model and constraints.

Wasabi.delete! — Function

delete(db::Any, model::T) where {T <: Model}
Deletes the given model from the database.

Wasabi.delete_all! — Function

delete_all(db::Any, m::Type{T}) where {T <: Model}
Deletes all rows from the given model.

Wasabi.delete_schema — Function

delete_schema(db::Any, m::Type{T}) where {T <: Model}
Deletes the schema for the given model.

Wasabi.df2model — Method

df2model(m::Type{T}, df::DataFrame) where {T <: Model}
Converts the given DataFrame to the given model.

Wasabi.disconnect — Function

disconnect(db::Any)
Disconnects from the database.

Wasabi.execute_query — Function

execute_query(db::Any, query::RawQuery, params::Vector{Any})
Executes the given query with the given parameters.

Wasabi.first — Function

first(db::Any, m::Type{T})
Returns the first row of the given model with the given id.

Wasabi.foreign_keys — Method

foreign_keys(m::Type{T})::Vector{ForeignKeyConstraint} where {T<:Model}
Returns the foreign key constraints for the given model.

Wasabi.from_sql_value — Function

from_sql_value(db::Type{Any}, v::Any)
Converts the SQL value to the struct type value.

Wasabi.insert! — Function

insert(db::Any, model::T) where {T <: Model}
Inserts the given model into the database.

Wasabi.isnullable — Method

isnullable(m::Type{T}, field::Symbol, constraints::Vector{S}) where {T <: Model,S<:Wasabi.ModelConstraint}
Returns true if the given column is nullable.

Wasabi.mapping — Function

mapping(db::Type{S}, t::Type{T})::String where {S,T}
Returns the mapping for the given database and type.

Wasabi.model2tuple — Method

model2tuple(m::T) where {T <: Model}
Converts the given model to a tuple.

Wasabi.primary_key — Method

primary_key(m::Type{T})::Union{PrimaryKeyConstraint, Nothing} where {T<:Model}
Returns the primary key constraint for the given model.

Wasabi.rollback — Function

rollback(db::Any)
Rolls back the current transaction.

Wasabi.tablename — Method

tablename(m::String)
Returns the name of the table for the given model.

Wasabi.tablename — Method

tablename(m::Type{T}) where {T <: Model}
Returns the name of the table for the given model.

Wasabi.to_sql_value — Function

to_sql_value(db::Type{Any}, v::Any)
Converts the given value to a value that can be used in a SQL query.

Wasabi.unique_constraints — Method

unique_constraints(m::Type{T})::Vector{UniqueConstraint} where {T<:Model}
Returns the unique constraints for the given model.

Wasabi.update! — Function

update(db::Any, model::T) where {T <: Model}
Updates the given model in the database.