QueryBuilder

You can use QueryBuilder to create and execute a query in a simple and fast way.

using Wasabi

mutable struct User <: Wasabi.Model
    id::Int
    name::String
end

configuration = Wasabi.SQLiteConnectionConfiguration("test.db")
conn = Wasabi.connect(configuration)

query = QueryBuilder.select(User) |> QueryBuilder.limit(1)
users = Wasabi.df2model(User, Wasabi.execute_query(conn, user))

Wasabi.QueryBuilder.Join

Wasabi.QueryBuilder.Query

Wasabi.QueryBuilder.groupby

Wasabi.QueryBuilder.join

Wasabi.QueryBuilder.limit

Wasabi.QueryBuilder.offset

Wasabi.QueryBuilder.orderby

Wasabi.QueryBuilder.select

Wasabi.QueryBuilder.Join — Type

Join{T<:Wasabi.Model}
Represents a join for the given model.

Wasabi.QueryBuilder.Query — Type

Query{T<:Wasabi.Model}
Represents a query for the given model.

Wasabi.QueryBuilder.groupby — Method

groupby(groupby::Vector{Symbol})
Sets the group by clause for the given query.

Wasabi.QueryBuilder.join — Method

join(m::Type{T}, type::Symbol, on::Tuple{Symbol, Symbol}, select::Vector{Symbol} = Symbol[]) where {T <: Model}
Joins the given model to the query.

Wasabi.QueryBuilder.limit — Method

limit(limit::Int)
Sets the limit for the given query.

Wasabi.QueryBuilder.offset — Method

offset(offset::Int)
Sets the offset for the given query.

Wasabi.QueryBuilder.orderby — Method

orderby(orderby::Vector{Symbol})
Sets the order by clause for the given query.

Wasabi.QueryBuilder.select — Method

select(source::Type{T}, select::Union{Vector{Symbol}, Nothing} = nothing)::Query where {T <: Model}
Creates a new query for the given model and selected columns. If no columns are specified, all columns are selected.